fix type

+++ interfaces/acados_template/acados_template/c_templates_tera/acados_solver.in.c
@@ -2209,7 +2209,7 @@ static void {{ model.name }}_acados_create_set_opts({{ model.name }}_solver_caps
     for (int i = 0; i < N; i++)
         ocp_nlp_solver_opts_set_at_stage(nlp_config, nlp_opts, i, "dynamics_newton_iter", &newton_iter_val);
 
-    int newton_tol_val = {{ solver_options.sim_method_newton_tol }};
+    double newton_tol_val = {{ solver_options.sim_method_newton_tol }};
     for (int i = 0; i < N; i++)
         ocp_nlp_solver_opts_set_at_stage(nlp_config, nlp_opts, i, "dynamics_newton_tol", &newton_tol_val);
